moved camel_to_space into util.py
