+++ util/__init__.py
@@ -1,0 +1,3 @@
+'''Useful classes and functions used elsewhere in the
+MuddySwamp engine'''
+from util.misc import *
